// frontend/src/types.ts
export type Role = 'ADMIN' | 'FARMER' | 'TRANSPORTER' | 'DISTRIBUTOR'
export type OrgType = 'FARMER_HOUSEHOLD' | 'COOPERATIVE' | 'COMPANY' | 'PROCESSOR' | 'DISTRIBUTOR' | 'RETAILER' | 'EXPORTER'
export type Species = 'SHRIMP' | 'CRAB'
export type BatchStatusCode = 'PREPARING' | 'GROWING' | 'READY_FOR_HARVEST' | 'HARVESTED' | 'COMPLETED' | 'CANCELLED'
export type BlockchainStatus = 'PENDING' | 'SUCCESS' | 'FAILED'

export interface Organization {
  id: string
  name: string
  type: OrgType
  address: string | null
  status: boolean
  created_at: string
}

export interface UserProfile {
  id: string
  organization_id: string | null
  full_name: string
  email: string | null
  phone: string | null
  role: Role
  wallet_address: string | null
  status: boolean
  created_at: string
  organization: Organization | null
}
// ...
export interface AuthResponse {
  success: boolean
  requiresConfirmation?: boolean
  accessToken: string | null
  user: UserProfile
}